Carrinho de Compras (Em Progresso)

+++ src/e_commerce/Model/Product.java
@@ -18,6 +18,15 @@ public class Product {
     private double price;
     private Integer quantity;
 
+    public Product() {
+    }
+
+    public Product(int id, String name, double price, int quantity) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+        this.quantity = quantity;
+    }
     public int getId() {
         return id;
     }
@@ -83,8 +92,23 @@ public class Product {
     private boolean productExists(int id) {
         Connect connect = new Connect();
         String sql = "SELECT COUNT(id) FROM Product WHERE id =" + id;
-        connect.executeSQL(sql);
-        return false;
+        ResultSet resultSet = connect.executeQuery(sql);
+        boolean check = false;
+
+        try {
+            if (resultSet.next()) {
+                int count = resultSet.getInt(1);
+                if (count != 0) {
+                    check = true;
+                } else {
+                    check = false;
+                }
+            }
+        } catch (SQLException ex) {
+            ex.printStackTrace();
+        }
+
+        return check;
     }
     public void createProduct() {
         Scanner scanner = new Scanner(in);
@@ -108,8 +132,6 @@ public class Product {
             var priceProduct = scanner.nextLine();
             out.print("Insira a nova quantidade do produto: ");
             var quantityProduct = scanner.nextLine();
-
-            // Realizar a atualização do produto
             updateProduct(Integer.parseInt(id), nameProduct, Double.parseDouble(priceProduct), Integer.parseInt(quantityProduct));
         } else {
             out.println("Produto não encontrado.");
@@ -120,8 +142,33 @@ public class Product {
         Scanner scanner = new Scanner(in);
         out.print("Insira o código do produto a ser removido: ");
         var id = scanner.nextLine();
-        this.deleteProduct( Integer.parseInt(id));
+        if (productExists(Integer.parseInt(id))) {
+            this.deleteProduct(Integer.parseInt(id));
+        } else {
+            out.println("Produto não encontrado.");
+        }
     }
+
+    public void showProducts() {
+        List<Product> productList = this.getProducts();
+
+        for (Product p : productList) {
+            out.println("#############################");
+            out.println("ID: " + p.getId());
+            out.println("Nome: " + p.getName());
+            out.println("Preço: " + p.getPrice());
+            out.println("Quantidade: " + p.getQuantity());
+            out.println("#############################\n");
+        }
+    }
+
+    public void findProduct() {
+        Scanner scanner = new Scanner(in);
+        out.print("Digite o nome do produto que quer encontrar: ");
+        var productName = scanner.nextLine();
+        this.showProductsByName(productName);
+    }
+
     public void registerProduct(String name, Double price, Integer quantity) {
         Connect connect = new Connect();
         String sql = "INSERT INTO Product (nameProduct, price, quantity) VALUES ('" + name + "','" + price + "','" + quantity + "')";
@@ -166,4 +213,25 @@ public class Product {
 
         return productList;
     }
+    public void showProductsByName(String productName) {
+        List<Product> productList = this.getProducts();
+        boolean productFound = false;
+
+        for (Product p : productList) {
+            if (p.getName().toLowerCase().contains(productName.toLowerCase())) {
+                out.println("#############################");
+                out.println("ID: " + p.getId());
+                out.println("Nome: " + p.getName());
+                out.println("Preço: " + p.getPrice());
+                out.println("Quantidade: " + p.getQuantity());
+                out.println("#############################\n");
+                productFound = true;
+            }
+        }
+
+        if (!productFound) {
+            out.println("Nenhum produto encontrado com o nome contendo '" + productName + "'.");
+        }
+    }
+
 }
